refactor: move SpinalSimStreamExtenderTester
